Fixed compiler warnings

+++ include/nscapi/nscapi_metrics_helper.cpp
@@ -23,12 +23,12 @@ namespace metrics {
 void add_metric(PB::Metrics::MetricsBundle *b, const std::string &key, long long value) {
   PB::Metrics::Metric *m = b->add_value();
   m->set_key(key);
-  m->mutable_gauge_value()->set_value(value);
+  m->mutable_gauge_value()->set_value(static_cast<double>(value));
 }
 void add_metric(PB::Metrics::MetricsBundle *b, const std::string &key, unsigned long long value) {
   PB::Metrics::Metric *m = b->add_value();
   m->set_key(key);
-  m->mutable_gauge_value()->set_value(value);
+  m->mutable_gauge_value()->set_value(static_cast<double>(value));
 }
 void add_metric(PB::Metrics::MetricsBundle *b, const std::string &key, std::string value) {
   PB::Metrics::Metric *m = b->add_value();
