+++ modules/CheckEventLog/realtime_thread.cpp
@@ -91,7 +91,7 @@ void real_time_thread::thread_proc() {
     if (dur && dur->total_milliseconds() < 0)
       dwWaitTime = 0;
     else if (dur)
-      dwWaitTime = dur->total_milliseconds();
+      dwWaitTime = static_cast<DWORD>(dur->total_milliseconds());
 
     NSC_DEBUG_MSG("Sleeping for: " + str::xtos(dwWaitTime) + "ms");
     DWORD dwWaitReason = WaitForMultipleObjects(static_cast<DWORD>(evlog_list.size() + 1), handles, FALSE, dwWaitTime);
